(#1226) Moved ForEach primitives to func package
